Add `runIsolated_` and change behavior of `runIsolated` (#24)
The old runIsolated function threw away the result value, but sometimes it's nice to have access to this. Here we rename this behavior to runIsolated_ and have runIsolated return the value. This is a minor breaking change, however the fix is simply to call runIsolated_ instead of runIsolated in your code.

I chose this breaking change over a less disruptive one to mimic the underscore pattern that is standard among monadic functions in the ecosystem, such as sequence_/sequence or mapM_/mapM.

Co-authored-by: nbloomf <nathan.bloomfield@automattic.com>