Remove the LogPrinterLock tasty option
We use an MVar as a lock to allow for printing log entries atomically. For some reason I had abused tasty's option system to acquire this MVar; I can't remember now what the reason for that was, but it doesn't make sense to me now. This patch removes the option and does the sensible thing instead.